Prefix all fixture cookbooks with `jenkins_*`
