+++ Berksfile
@@ -1,15 +1,20 @@
+def fixture(name)
+  cookbook "jenkins_#{name}", path: "test/fixtures/cookbooks/jenkins_#{name}"
+end
+
 source 'https://supermarket.chef.io'
 
 metadata
 
 group :integration do
-  cookbook 'authentication',      path: 'test/fixtures/cookbooks/authentication'
-  cookbook 'jenkins_command',     path: 'test/fixtures/cookbooks/jenkins_command'
-  cookbook 'jenkins_credentials', path: 'test/fixtures/cookbooks/jenkins_credentials'
-  cookbook 'jenkins_job',         path: 'test/fixtures/cookbooks/jenkins_job'
-  cookbook 'jenkins_plugin',      path: 'test/fixtures/cookbooks/jenkins_plugin'
-  cookbook 'jenkins_script',      path: 'test/fixtures/cookbooks/jenkins_script'
-  cookbook 'jenkins_slave',       path: 'test/fixtures/cookbooks/jenkins_slave'
-  cookbook 'jenkins_user',        path: 'test/fixtures/cookbooks/jenkins_user'
-  cookbook 'smoke',               path: 'test/fixtures/cookbooks/smoke'
+  cookbook 'chef-sugar'
+  fixture 'authentication'
+  fixture 'command'
+  fixture 'credentials'
+  fixture 'job'
+  fixture 'plugin'
+  fixture 'script'
+  fixture 'slave'
+  fixture 'smoke'
+  fixture 'user'
 end
