--- test/fixtures/cookbooks/authentication/metadata.rb
@@ -1,2 +1,0 @@
-name 'authentication'
-depends 'jenkins'
