+++ test/fixtures/cookbooks/jenkins_authentication/metadata.rb
@@ -1,0 +1,2 @@
+name 'jenkins_authentication'
+depends 'jenkins_server_wrapper'
