+++ test/fixtures/cookbooks/jenkins_authentication/recipes/private_key.rb
@@ -1,25 +1,27 @@
-
-# PAM auth acts weird when run under runit
-node.set['jenkins']['master']['install_method'] = 'package'
-# Install the latest version
-node.set['jenkins']['master']['version']        = nil
-include_recipe 'jenkins::master'
+include_recipe 'jenkins_server_wrapper::default'
 
 #
 # Setup
 # ------------------------------
 
-# Ensure the `jenkins` user is part of the `shadow` group
-group 'shadow' do
-  append true
-  members %w(jenkins)
-  notifies :restart, 'service[jenkins]', :immediately
-end
+# Load the private key from a data bag item. This should be an encrypted data
+# bag item in real deployments.
+jenkins = data_bag_item('keys', 'jenkins')
 
-# Set the username/password on the executor. You should pull these from
-# a secret store or encrypted data bag item.
-node.run_state[:jenkins_username] = 'vagrant'
-node.run_state[:jenkins_password] = 'vagrant'
+require 'openssl'
+require 'net/ssh'
+key = OpenSSL::PKey::RSA.new(jenkins['private_key'])
+private_key = key.to_pem
+public_key  = "#{key.ssh_type} #{[key.to_blob].pack('m0')}"
+
+# Set the private key on the executor
+node.run_state[:jenkins_private_key] = private_key
+
+# Create a default Chef user with the public key
+jenkins_user 'chef' do
+  full_name   'Chef Client'
+  public_keys [public_key]
+end
 
 # Turn on basic authentication
 jenkins_script 'setup authentication' do
@@ -28,7 +30,7 @@ jenkins_script 'setup authentication' do
     def instance = Jenkins.getInstance()
 
     import hudson.security.*
-    def realm = new PAMSecurityRealm(null)
+    def realm = new HudsonPrivateSecurityRealm(false)
     instance.setSecurityRealm(realm)
 
     def strategy = new hudson.security.FullControlOnceLoggedInAuthorizationStrategy()
