+++ test/fixtures/cookbooks/jenkins_authentication/recipes/username_password.rb
@@ -1,28 +1,20 @@
-
-include_recipe 'jenkins::master'
+include_recipe 'jenkins_server_wrapper::default'
 
 #
 # Setup
 # ------------------------------
 
-# Load the private key from a data bag item. This should be an encrypted data
-# bag item in real deployments.
-jenkins = data_bag_item('keys', 'jenkins')
-
-require 'openssl'
-require 'net/ssh'
-key = OpenSSL::PKey::RSA.new(jenkins['private_key'])
-private_key = key.to_pem
-public_key  = "#{key.ssh_type} #{[key.to_blob].pack('m0')}"
-
-# Set the private key on the executor
-node.run_state[:jenkins_private_key] = private_key
-
-# Create a default Chef user with the public key
-jenkins_user 'chef' do
-  full_name   'Chef Client'
-  public_keys [public_key]
+# Ensure the `jenkins` user is part of the `shadow` group
+group 'shadow' do
+  append true
+  members %w(jenkins)
+  notifies :restart, 'service[jenkins]', :immediately
 end
+
+# Set the username/password on the executor. You should pull these from
+# a secret store or encrypted data bag item.
+node.run_state[:jenkins_username] = 'vagrant'
+node.run_state[:jenkins_password] = 'vagrant'
 
 # Turn on basic authentication
 jenkins_script 'setup authentication' do
@@ -31,7 +23,7 @@ jenkins_script 'setup authentication' do
     def instance = Jenkins.getInstance()
 
     import hudson.security.*
-    def realm = new HudsonPrivateSecurityRealm(false)
+    def realm = new PAMSecurityRealm(null)
     instance.setSecurityRealm(realm)
 
     def strategy = new hudson.security.FullControlOnceLoggedInAuthorizationStrategy()
