+++ test/fixtures/cookbooks/jenkins_smoke/metadata.rb
@@ -1,9 +1,10 @@
-name 'smoke'
-depends 'authentication'
+name 'jenkins_smoke'
+depends 'jenkins_authentication'
 depends 'jenkins_command'
 depends 'jenkins_credentials'
 depends 'jenkins_job'
 depends 'jenkins_plugin'
 depends 'jenkins_script'
+depends 'jenkins_server_wrapper'
 depends 'jenkins_slave'
 depends 'jenkins_user'
